refactor sam/vam function package
This commit refactors the function packages so you can add a new
function with a vector only implementation.  There is more work here
to clean things up further but this gets us moving forward for now.
