+++ compiler/semantic/expr.go
@@ -15,8 +15,8 @@ import (
 	"github.com/brimdata/super/pkg/nano"
 	"github.com/brimdata/super/pkg/reglob"
 	"github.com/brimdata/super/runtime/sam/expr"
-	"github.com/brimdata/super/runtime/sam/expr/function"
 	"github.com/brimdata/super/runtime/vam/expr/agg"
+	"github.com/brimdata/super/runtime/vam/expr/function"
 	"github.com/brimdata/super/sup"
 	"github.com/shellyln/go-sql-like-expr/likeexpr"
 )
